Gooseable single slot paxos replica; clerk is unimplemented

+++ paxi/single/common.go
@@ -2,5 +2,5 @@ package single
 
 type ValType = uint64
 
-const PREPARE = 1
-const PROPOSE = 2
+const PREPARE = uint64(1)
+const PROPOSE = uint64(2)
